Supprime le sous-composant "ResponsableSelectionnable"
... car le `bind:group` ne fonctionne pas dans les sous composant, cf https://github.com/sveltejs/svelte/issues/2308
